lpm-levels: Allow enforcing deepest cluster idle state selection
Similarly to the use_deepest_state flag added to cpuidle in commit
d29563d2fd4f ("cpuidle: Allow enforcing deepest idle state selection"),
add a cluster_use_deepest_state flag to lpm-levels that forces the
cluster idle state to be the deepest possible state at the time of
selection. Note that this is *not* always the deepest state available as
the status of RPM will influence whether the system is able to enter the
deepest state(s).

This flag can be modified from within the kernel as well as from
userspace through a module parameter.

The deepest cluster idle state available on sdm845 uses 76% less power
than the shallowest idle state.

Change-Id: I8f222e5b87d92f69d5881699a341dcd5afe6934f
Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: prorooter007 <shreyashwasnik112@gmail.com>