arm: dts: msm8937: Remove system mount partition
As of this commit 143322b12335354e4571e02d2a778eda6363c5dc
system partition has been removed for non A/B devices which is required for system-as-root in Android 9.
A/B devices do not need to change as they already must use system-as-root.

Read this for more info https://source.android.com/devices/bootloader/system-as-root

Signed-off-by: Hemant Sharma <hemantbeast@gmail.com>